feat: replace login component and register component >> add to app pages
